requirements.txt
ccxt==4.2.85
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0
TA-Lib==0.4.28
python-dateutil>=2.8.2